feat: add doctor details endpoint and refactor appointments endpoints
